+++ src/flake8_custom_import_rules/defaults.py
@@ -10,7 +10,7 @@ else:
 class Settings(TypedDict, total=False):
     """The default settings for the flake8_custom_import_rules plugin."""
 
-    base_package: str
+    base_package: str | list[str]
     # version: Optional[str]
     # version_files: List[str]
     # version_provider: Optional[str]